Fix native compile and test benchmarks. A bit more code coverage.
