+++ src/commonTest/kotlin/electionguard/decrypt/RunDecryptingMediatorTest.kt
@@ -1,4 +1,4 @@
-package electionguard.workflow
+package electionguard.decrypt
 
 import electionguard.core.productionGroup
 import electionguard.decrypt.readDecryptingTrustees
@@ -13,16 +13,23 @@ class RunDecryptingMediatorTest {
         val group = productionGroup()
         val inputDir = "src/commonTest/data/runWorkflowAllAvailable"
         val trusteeDir = "src/commonTest/data/runWorkflowAllAvailable/private_data/trustees"
-        val outputDir = "testOut/testDecryptingMediator"
+        val outputDir = "testOut/testDecryptingMediatorAll"
         runDecryptTally(group, inputDir, outputDir, readDecryptingTrustees(group, inputDir, trusteeDir), "createdBy")
     }
 
     @Test
     fun testDecryptingMediatorSome() {
-        val group = productionGroup()
-        val inputDir = "src/commonTest/data/runWorkflowSomeAvailable"
-        val trusteeDir = "src/commonTest/data/runWorkflowSomeAvailable/private_data/trustees"
-        val outputDir = "testOut/testDecryptingMediator"
-        runDecryptTally(group, inputDir, outputDir, readDecryptingTrustees(group, inputDir, trusteeDir), "createdBy")
+        main(
+            arrayOf(
+                "-in",
+                "src/commonTest/data/runWorkflowSomeAvailable",
+                "-trustees",
+                "src/commonTest/data/runWorkflowSomeAvailable/private_data/trustees",
+                "-out",
+                "testOut/testDecryptingMediatorSome",
+                "-createdBy",
+                "testDecryptingMediatorSome"
+            )
+        )
     }
 }
