+++ src/commonTest/kotlin/electionguard/keyceremony/RunKeyCeremonyTest.kt
@@ -1,0 +1,22 @@
+package electionguard.keyceremony
+
+import kotlin.test.Test
+
+class RunKeyCeremonyTest {
+
+    @Test
+    fun testKeyCeremonyMain() {
+        main(
+            arrayOf(
+                "-in",
+                "src/commonTest/data/start",
+                "-trustees",
+                "testOut/RunKeyCeremonyTest/private_data/trustees",
+                "-out",
+                "testOut/RunKeyCeremonyTest",
+            )
+        )
+    }
+
+}
+
