--- src/nativeMain/kotlin/PerfTest.kt
@@ -1,5 +1,0 @@
-import electionguard.core.simpleBallotPerfTest
-
-fun main() {
-    simpleBallotPerfTest()
-}
